requirements.txt
Pillow
pytesseract==0.3.8